+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 0.0.14
+  - Clean code
   - Additional error & exception handling
   - Add the support for CreateFulfillmentOrder action
   - Added typeScript example of reports
+  - Added javaScript example of Filter Orders
+  - Added javaScript example of override/set the default options before making the request.
 
 0.0.13 December 08, 2017
   - Add the MWS Version Detail